requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
twilio==9.10.4
pydantic==2.12.5
python-dotenv==1.2.2
groq==1.1.1
sqlalchemy==2.0.48
python-multipart==0.0.22